chore: rewritten double-to-string algo that prints to highest possible precision.

+++ core/src/main/java/io/questdb/griffin/engine/functions/cast/CastByteToLongFunctionFactory.java
@@ -49,7 +49,7 @@ public class CastByteToLongFunctionFactory implements FunctionFactory {
 
         @Override
         public long getLong(Record rec) {
-            return arg.getLong(rec);
+            return arg.getByte(rec);
         }
     }
 }
